[New Sample] Create Mobile Geodatabase * Converted to MVC layout
