+++ feature_layers/create-mobile-geodatabase/build.gradle
@@ -11,7 +11,7 @@ ext {
 
 javafx {
     version = "17.0.2"
-    modules = [ 'javafx.controls' ]
+    modules = [ 'javafx.controls', 'javafx.fxml' ]
 }
 
 compileJava.options.encoding = 'UTF-8'
