+++ feature_layers/create-mobile-geodatabase/src/main/resources/create_mobile_geodatabase/style.css
@@ -1,0 +1,3 @@
+.panel-region .label {
+  -fx-text-fill: white;
+}
